requirements.txt
streamlit
crewai
autogen
Langchain
python-dotenv
googlesearch-python
openai
requests
chromadb